collection view: optimize selections
we now use a delegate to detect selections, and that behavior has been
moved to the view controller, aside from the view background color
modification logic. this way, cmd+a doesn't lag as it tries to fetch all
tags (and wastes them in the process) and all of the tag fetching stuff
resides in the view controller.