Remove custom drawing for grid thumbnails
Previously, grid thumbnails were drawn with a custom NSView subclass,
SelectableImageView. This class was problematic for a number of reasons:

- In order to tint selected images, a copy of the thumbnail is made and
  custom drawing is performed. This seems unnecessary; instead, one
  could draw *over* the image.
- We're reimplementing a lot of what NSBox already provides (drawing a
  custom background color and border).
- We're setting the layer contents to an NSImage, which is inherently
  troublesome.

Remove this custom view and draw thumbnails with standard NSImageViews
inside of NSBoxes. Unfortunately, NSImageView doesn't provide any
standard machinery for proportionally scaling the contained image to
fill the view's frame, so I've removed the "grid thumbnail scaling"
preference for now.

Furthermore, selected thumbnails are no longer tinted, but I think this
is fine, because the images are laid out in such a way that it's easy to
discern whether they are selected or not.