Remove NSVisualEffectView behind the posts grid
It doesn't provide much utility: we want the thumbnails to be as visible
as possible as they're one of the most important UI elements on screen,
and having a blurred background behind them doesn't help. It also
(arguably) reduces contrast and makes it harder to see the thumbnails in
the first place.