Only append new posts when updating the snapshot
When loading more posts, we update the existing snapshot instead of
creating a new one. However, we indiscriminately add all of the posts
within the latest listing chunk to the new snapshot. This isn't a
problem if the latest chunk contains new posts, but if it doesn't (such
as when the listing is exhausted), we apply with duplicate item
identifiers.

This results in a warning, and more fatally, the selection mysteriously
being reset if it happened to be nearby.

Fix this by only ever appending new posts that aren't already in the
snapshot, and don't bother applying at all if there are no new posts.