Wrap around when using horizontal arrow keys
When navigating the posts grid with the arrow keys, the left and right
keys don't wrap around to the next (or previous) row when hitting the
edge of the column. This is annoying, so fix that.

Cocoa isn't going to help us here, so reimplementing the behavior from
scratch was necessary.